+++ src/resources/views/inc/sidebar.blade.php
@@ -12,8 +12,8 @@
           <!-- ================================================ -->
           <!-- ==== Recommended place for admin menu items ==== -->
           <!-- ================================================ -->
-          <li><a href="{{ backpack_url('dashboard') }}"><i class="fa fa-dashboard"></i> <span>{{ trans('backpack::base.dashboard') }}</span></a></li>
 
+          @include('backpack::inc.sidebar_content')
 
           <!-- ======================================= -->
           {{-- <li class="header">Other menus</li> --}}