+++ CHANGELOG.md
@@ -21,9 +21,11 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Security
 - Nothing
 
-------
+-------
 
 ### Added
+- flexbox css helper class;
+- support for HTML messages inside Alert bubbles, when triggered from PHP;
 - added publish lang command;
 - added command to publish only the minimum amount of files needed for Backpack to work;
 - ```sidebar_content.blade.php``` file, so that we can add sidebar items using a command